requirements.txt
boto==2.38.0
cffi==1.1.2
cryptography==0.9.1
Django==1.8.2
django-bootstrap3==6.1.0
django-ckeditor==4.4.8
django-meta==0.3.1
django-robots==1.1
django-storages==1.1.8
django-subdomains==2.0.4
django-suit==0.2.13
docutils==0.12
enum34==1.0.4
gitchangelog==2.2.1
idna==2.0
ipaddress==1.0.7
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
Pillow==2.9.0
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
pytz==2015.4
requests==2.7.0
six==1.9.0
